+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# 1.0.0 --installs on--> 3.1 IDE
+# 1.0.0 --installs on--> 2.2.4 IDE-EE / 3.1 IDE-CE
 
 ## Miscellaneous
 
 - [FP-2027 - Editor shouldn't require method "renderRightMenu" implementation](https://movai.atlassian.net/browse/FP-2027)
+- [FP-2036 - Fix data type validation](https://movai.atlassian.net/browse/FP-2036)
+  - [Fix Boolean type validation : Accept js boolean](https://github.com/MOV-AI/frontend-npm-lib-ide/commit/0fe2d8e4d121fe702bfcfd9db1a1ac1702d35625)
+  - [Implement ConfigurationSelector for Configuration Data Type](https://github.com/MOV-AI/frontend-npm-lib-ide/commit/0f063cd3c627e1f0ebd5a0d30967be9cc95931a1)
+  - [Fix Configuration Type : Works with parameter syntax and default one](https://github.com/MOV-AI/frontend-npm-lib-ide/commit/4e1502c1724c6ff0d5a8cc327b61fd70f88c77cb)